GEODE-9190 Cleanup of Stats Code Warnings (#6347)
Cleanup of Stats Warnings

+++ geode-core/src/distributedTest/java/org/apache/geode/ClusterCommunicationsDUnitTest.java
@@ -326,7 +326,7 @@ public class ClusterCommunicationsDUnitTest implements Serializable {
     memberVM.invoke("perform update", () -> {
       DMStats stats = ((InternalDistributedSystem) cache.getDistributedSystem())
           .getDistributionManager().getStats();
-      int reconnectAttempts = stats.getReconnectAttempts();
+      long reconnectAttempts = stats.getReconnectAttempts();
       cache.getRegion(regionName).put("testKey", "updatedTestValue");
       assertThat(stats.getReconnectAttempts()).isEqualTo(reconnectAttempts);
     });
