+++ geode-core/src/distributedTest/java/org/apache/geode/cache30/SlowRecDUnitTest.java
@@ -604,7 +604,7 @@ public class SlowRecDUnitTest extends JUnit4CacheTestCase {
     // updates to a non-conflating are not.
     LogWriterUtils.getLogWriter().info("[testConflationSequence] conflate & no-conflate regions");
     forceQueuing(r);
-    final int initialAsyncSocketWrites = stats.getAsyncSocketWrites();
+    final long initialAsyncSocketWrites = stats.getAsyncSocketWrites();
 
     value = "count=" + count;
     lastValue = value;
@@ -1228,7 +1228,7 @@ public class SlowRecDUnitTest extends JUnit4CacheTestCase {
     // set others before vm0 connects
     final Set others = dm.getOtherDistributionManagerIds();
     long initialQueuedMsgs = stats.getAsyncQueuedMsgs();
-    final int initialQueues = stats.getAsyncQueues();
+    final long initialQueues = stats.getAsyncQueues();
 
     // create receiver in vm0 with queuing enabled
     final Properties p = new Properties();
